Amount of item as production goal can be a fraction of an Recipe time in seconds is not always an integer. #lexmall

+++ src/main/java/org/joelson/mattias/foundry/resource_calculator/model/ProductionGraph.java
@@ -30,13 +30,13 @@ public class ProductionGraph {
         this.productionTableItemNameRows = Objects.requireNonNull(productionTableItemNameRows);
     }
 
-    public void addGoals(Map<Item, Integer> productionGoals) {
-        for (Map.Entry<Item, Integer> goalEntry : productionGoals.entrySet()) {
+    public void addGoals(Map<Item, Float> productionGoals) {
+        for (Map.Entry<Item, Float> goalEntry : productionGoals.entrySet()) {
             addItemsPerMinute(goalEntry.getKey(), goalEntry.getValue());
         }
     }
 
-    private void addItemsPerMinute(Item item, Integer itemsPerMinute) {
+    private void addItemsPerMinute(Item item, Float itemsPerMinute) {
         itemNodeMap.get(item.name()).addItemsPerMinute(itemsPerMinute);
     }
